requirements.txt
streamlit==1.42.0
python-docx==1.1.2
pymupdf==1.25.1
pdfplumber==0.11.4
spacy==3.8.2
nltk==3.9.1
langdetect==1.0.9
groq==0.4.2           <-- ADD THIS (Crucial)
python-dotenv==1.0.1
pandas==2.2.3
fpdf2==2.7.8